inicio do client java fx
