+++ rest-api5-client-componente3/src/main/java/com/upt/lp/app/JavaFxLauncher.java
@@ -1,0 +1,10 @@
+package com.upt.lp.app;
+
+import javafx.application.Application;
+
+public class JavaFxLauncher {
+
+    public static void main(String[] args) {
+        Application.launch(EquipamentoClient.class, args);
+    }
+}
